improvement: change validate from an assertion function to a type guard
Usage of parser.validate is unclear because you need to assign an output type, which defeats the
purpose of this whole library (see https://github.com/microsoft/TypeScript/issues/34596 and related
issues.) This changes .validate (which throws on invalid input) to .validates, which is a type guard
that returns false on invalid input. .parse still throws on invalid input.

BREAKING CHANGE: TsjsonParser.validate renamed to TsjsonParser.validates and functionality changed
from assertion function to type guard.